requirements.txt
# Core ML/DL packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=1.0.0

# Computer Vision and Image Processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data Science and Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0.0

# Optional but recommended for CUDA support
# Uncomment if you have CUDA-compatible GPU:
# torch-audio
# torch-vision with CUDA support
